9. Numbers have life; they're not just symbols on paper

+++ 0x06-pointers_arrays_strings/101-print_number.c
@@ -7,7 +7,7 @@
 
 void print_number(int n)
 {
-	long num, t, i, size, power;
+	unsigned int num, t, i, size, power;
 
 	size = 0;
 	power = 1;
